fix - TurnService 적용
